+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/RunLQNS.java
@@ -1,9 +1,9 @@
 package de.fabiankeller.palladio;
 
 import de.fabiankeller.palladio.analysis.PcmProvider;
+import de.fabiankeller.palladio.analysis.pcm2lqn.runner.Pcm2LqnAnalysisConfig;
+import de.fabiankeller.palladio.analysis.pcm2lqn.runner.Pcm2LqnRunner;
 import de.fabiankeller.palladio.analysis.provider.FileSystemProvider;
-import de.fabiankeller.palladio.analysis.runner.pcm2lqn.Pcm2LqnAnalysisConfig;
-import de.fabiankeller.palladio.analysis.runner.pcm2lqn.Pcm2LqnRunner;
 import de.fabiankeller.palladio.config.PcmModelConfig;
 import de.fabiankeller.palladio.environment.PalladioEclipseEnvironment;
