+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/RunLqnsWithBuilder.java
@@ -1,8 +1,8 @@
 package de.fabiankeller.palladio;
 
+import de.fabiankeller.palladio.analysis.pcm2lqn.runner.Pcm2LqnAnalysisConfig;
+import de.fabiankeller.palladio.analysis.pcm2lqn.runner.Pcm2LqnRunner;
 import de.fabiankeller.palladio.analysis.provider.SimpleTacticsProvider;
-import de.fabiankeller.palladio.analysis.runner.pcm2lqn.Pcm2LqnAnalysisConfig;
-import de.fabiankeller.palladio.analysis.runner.pcm2lqn.Pcm2LqnRunner;
 import de.fabiankeller.palladio.builder.PcmBuilder;
 import de.fabiankeller.palladio.config.PcmModelConfig;
 import de.fabiankeller.palladio.environment.PalladioEclipseEnvironment;
